move warp load test logic to simulator dir
